% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{annotateCellTypesGSEA}
\alias{annotateCellTypesGSEA}
\title{Match deconvolved cell-types to ground truth cell-types based on transcriptional profiles}
\usage{
annotateCellTypesGSEA(beta, gset, qval = 0.05)
}
\arguments{
\item{beta}{cell-type (rows) x gene (columns) matrix of deconvolved cell-type transcriptional profiles}

\item{gset}{named list where each entry is a vector of marker genes for a given ground truth cell-type.}

\item{qval}{adjusted p-value threshold (default: 0.05)}
}
\value{
A list that contains
\itemize{
\item results: A named list that contains sorted matrices for each deconvolved cell-type.
The matrix rows are the ground truth cell-types ordered by significance, edge-score, and enrichment score
of their gene sets in the deconvolved transcriptional profile of a given deconvolved cell-type.
\item predictions: a named vector where the names are the deconvolved cell-types and the values
are the best matched ground truth cell-type that is also positively enriched.
}
}
\description{
Match deconvolved cell-types to ground truth cell-types by testing for
enrichment of ground truth marker gene sets in the deconvolved transcriptional profiles.
Uses \code{liger::iterative.bulk.gsea}.
}
