% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BCL}
\alias{BCL}
\title{Spatial transcriptomics transcriptomics of 4 breast cancer biopsy sections}
\format{
List where 'counts' is a sparse matrix with columns as voxels and rows as genes and
                   'pos' is a data frame of x and y position values per voxel
                         and slice index for 4 consecutive slices
}
\source{
\url{https://science.sciencemag.org/content/353/6294/78}
}
\usage{
BCL
}
\description{
Spatial transcriptomics transcriptomics of 4 breast cancer biopsy sections
}
\keyword{datasets}
