% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{bh.adjust}
\alias{bh.adjust}
\title{Benjamini-Hochberg p-value adjustment}
\usage{
bh.adjust(x, log = FALSE)
}
\arguments{
\item{x}{vector of p-values}

\item{log}{log option (default: FALSE)}
}
\description{
Benjamini-Hochberg p-value adjustment
}
