% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{buildBregmaCorpus}
\alias{buildBregmaCorpus}
\title{Generate simulated corpus as input into `topicmodels` and the matched ground
truth spot-topic proportion and topic-gene proportion matrices for a given
simulated bregma spot dataset (built with `simulateBregmaSpots`)}
\usage{
buildBregmaCorpus(hashTable, bregmaID)
}
\arguments{
\item{hashTable}{output hash table of simulated bregma spots from
`simulateBregmaSpots`}

\item{bregmaID}{ID to reference a given bregma in the hashTable. ex: "-0.04"}
}
\value{
a list that contains
\itemize{
\item sim: slam::as.simple_triplet_matrix(corpus); required format for topicmodels::LDA input
\item gtSpotTopics: the ground truth of major cell class proportions in each simulated patch.
    Use as ground truth "theta" (document x topix proportions)
\item gtCtGenes: the normalized gene count proportions of each of the major cell classes.
    Use as ground truth "beta" (topics x word frequencies)
\item cellCounts: data.frame that has patch names, "x", and "y" centroid coordinates,
    and counts of total cells in each patch
\item classColors: vector of colors for each cell class
\item annotDf: data.frame of the individual cells of the bregma,
    their position, cell type label, and assigned patch
}
}
\description{
Important: "Cell_class" column in "bregmaFullDf" as input for `simulateBregmaSpots`
    is where the cell labels will be pulled from. These will be used to construct
    the: "cellTypeTable", where downstream in `buildBregmaCorpus`, will be used
    to make "gtSpotTopics" and "gtCtGenes".
}
