% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clusterTopics}
\alias{clusterTopics}
\title{Cluster topics together using dynamic tree cutting.}
\usage{
clusterTopics(
  beta,
  clustering = "ward.D",
  dynamic = "hybrid",
  deepSplit = 4,
  plot = TRUE
)
}
\arguments{
\item{beta}{Beta matrix (topic-word distributions)}

\item{clustering}{Clustering agglomeration method to be used (default: ward.D)}

\item{dynamic}{Dynamic tree cutting method to be used (default: hybrid)}

\item{deepSplit}{Dynamic tree cutting sensitivity parameter (default: 4)}

\item{plot}{Boolean for plotting}

\item{distance}{Distance measure to be used (default: euclidean)}
}
\value{
A list that contains
\itemize{
\item order = vector of the dendrogram index order for the topics
\item clusters = factor of the topics (names) and their assigned cluster (levels)
\item dendro = dendrogram of the clusters
}
}
\description{
Cluster topics together using dynamic tree cutting.
}
