% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{correlationPlot_2}
\alias{correlationPlot_2}
\title{Generate heatmap of correlations}
\usage{
correlationPlot_2(
  mat,
  rowLabs = NA,
  colLabs = NA,
  rowv = NA,
  colv = NA,
  margins = c(6, 8),
  textSize = 0.9
)
}
\arguments{
\item{mat}{matrix with correlation values from -1 to 1}

\item{rowLabs}{y-axis label for plot. These are the rows of the matrix, or specifically m1 from getCorrMtx. (default: NULL)}

\item{colLabs}{x-axis label for plot. These are the columns of the matrix, or specifically m2 from getCorrMtx. (default: NULL)}

\item{rowv}{cluster order for the rows to make dendrogram (RowV in heatmap.2()). (default: NA)}

\item{colv}{cluster order for the columns to make dendrogram (Colv in heatmap.2()). (default: NA)}

\item{margins}{set margins of the plot. (default: c(6,8))}

\item{textSize}{set the text size for both x-axis and y-axis labels. (default: 0.9)}
}
\description{
Visualize the correlations between topics stored in a matrix, typically one
returned via \code{getCorrMtx()}. This function uses gplots::heatmap.2.
}
