% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fitLDA}
\alias{fitLDA}
\title{Find the optimal number of topics K for topic modeling}
\usage{
fitLDA(
  counts,
  Ks = seq(2, 10, by = 2),
  seed = 0,
  ncores = parallel::detectCores(logical = TRUE) - 1,
  plot = TRUE
)
}
\arguments{
\item{counts}{Gene expression counts with genes as rows}

\item{Ks}{vector of K parameters to search}

\item{seed}{Random seed}

\item{ncores}{Number of cores for parallelization}
}
\value{
A list that contains
\itemize{
\item models: each fitted LDA model
\item kOpt1: the optimal K based on Kneed algorithm
\item kOpt2: the optimal K based on minimum
\item perplexities: perplexity scores for each model
\item corpus: the corpus that was used to fit each model
}
}
\description{
Find the optimal number of topics K for topic modeling
}
