% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getBetaTheta}
\alias{getBetaTheta}
\title{Pull out topics and terms from fitted topic models from fitLDA
Jean: Brendan, the use of theta and beta vs topics and terms
across different functions is quite confusing
Please update for consistency}
\usage{
getBetaTheta(lda)
}
\description{
Pull out topics and terms from fitted topic models from fitLDA
Jean: Brendan, the use of theta and beta vs topics and terms
across different functions is quite confusing
Please update for consistency
}
