% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{getCorrMtx}
\alias{getCorrMtx}
\title{Find Pearson correlations between topics (cell types) with respect to their
proportions across documents (spots), i.e. thetas, or gene probabilities,
i.e. betas.}
\usage{
getCorrMtx(m1, m2, type, thresh = NULL)
}
\arguments{
\item{m1}{first matrix}

\item{m2}{second matrix}

\item{type}{must be either "t" (theta) or "b" (beta)}

\item{thresh}{if comparing betas, use to compare genes above this probability.
NULL or 0 < numeric < 1.0 (default: NULL)}
}
\value{
matrix of pearson correlations between topics m1 (rows) by topics m2 (cols)
}
\description{
Find Pearson correlations between topics (cell types) with respect to their
proportions across documents (spots), i.e. thetas, or gene probabilities,
i.e. betas.
}
