% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mOB}
\alias{mOB}
\title{Spatial transcriptomics transcriptomics of the mouse olfactory bulb}
\format{
List where 'counts' is a sparse matrix with columns as voxels and rows as genes and
'pos' is a data frame of x and y position values per voxel
}
\source{
\url{https://science.sciencemag.org/content/353/6294/78}
}
\usage{
mOB
}
\description{
Spatial transcriptomics transcriptomics of the mouse olfactory bulb
}
\keyword{datasets}
