% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mPOA}
\alias{mPOA}
\title{MERFISH data of the mouse pre-optic region for a female naive animal (FN7)}
\format{
List where 'mat' is a sparse matrix with columns as cells and rows as genes
                         where expression values have already been normalized by volume
                   'pos' is a data frame of x, y, z position values per cell
                         and brain position as 6 slice indices from anterior to posterior
}
\source{
\url{https://science.sciencemag.org/content/362/6416/eaau5324/}
}
\usage{
mPOA
}
\description{
MERFISH data of the mouse pre-optic region for a female naive animal (FN7)
}
\keyword{datasets}
