% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mtx2ggplotDf}
\alias{mtx2ggplotDf}
\title{Convert a matrix to a long form data frame for easy plotting via ggplot2
Jean: Brendan, please check if this function is necessary
or could be replaced with reshape2::melt}
\usage{
mtx2ggplotDf(mtx, colLab = "columns", rowLab = "rows", cellLab = "cells")
}
\arguments{
\item{mtx}{the matrix to be converted}

\item{colLab}{df column label for the mtx columns}

\item{rowLab}{df row label for the row labels}

\item{cellLab}{label for the cell values}
}
\description{
Convert a matrix to a long form data frame for easy plotting via ggplot2
Jean: Brendan, please check if this function is necessary
or could be replaced with reshape2::melt
}
