% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{perplexityPlot}
\alias{perplexityPlot}
\title{Plot the perplexity and rare cell-types versus fitted Ks}
\usage{
perplexityPlot(models, corpus = NULL, perc.rare.thresh = 0.05)
}
\arguments{
\item{models}{list returned from fitLDA}

\item{corpus}{If corpus is NULL, then it will use the original corpus that
the model was fitted to. Otherwise, compute deconvolved topics from this
new corpus. Needs to be pixels x genes and nonnegative integer counts.
Each row needs at least 1 nonzero entry (default: NULL)}

\item{perc.rare.thresh}{the number of deconvolved cell-types with mean pixel proportion below this fraction used to assess
performance of fitted models for each K. Recorded for each K. (default: 0.05)}
}
\value{
a plot indicating the perplexity and number of rare cell-types of a list of fitted LDA models
}
\description{
the same plot returned by fitLDA() but now callable as a
separate function.
}
\examples{
data(mOB)
pos <- mOB$pos
cd <- mOB$counts
counts <- cleanCounts(cd, min.lib.size = 100)
corpus <- restrictCorpus(counts, removeAbove=1.0, removeBelow = 0.05)
ldas <- fitLDA(t(as.matrix(corpus)), Ks = seq(2,6))
perplexityPlot(models = ldas, corpus = corpus)

}
