% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{preprocess}
\alias{preprocess}
\title{Pre-process ST gene count matrices to construct corpus for input into LDA}
\usage{
preprocess(
  dat,
  alignFile = NA,
  extractPos = FALSE,
  selected.genes = NA,
  nTopGenes = NA,
  genes.to.remove = NA,
  perc.spots = NA,
  min.reads = 100,
  min.lib.size = 100,
  min.detected = 1,
  ODgenes = TRUE,
  nTopOD = NA,
  od.genes.alpha = 0.05,
  gam.k = 5
)
}
\arguments{
\item{dat}{pixel (row) x gene (columns) mtx with gene counts OR path to it}

\item{alignFile}{path to 3x3 alignment file to adjust pixel coordinates
(optional).
Stahl 2016 datasets (default: NA)}

\item{extractPos}{Boolean to extract pixel positional coordinates from pixel name
names
(default: FALSE)}

\item{selected.genes}{vector of gene names to use specifically for the corpus
(default: NA)}

\item{nTopGenes}{integer for number of top expressed genes to remove
(default: NA)}

\item{genes.to.remove}{vector of gene names or patterns for matching to genes
to remove (default: NA). ex: c("^mt-") or c("^MT", "^RPL", "^MRPL")}

\item{perc.spots}{non-negative numeric <=1 to use as a percentage.
Removes genes present in this fraction of pixels (default: NA)}

\item{min.reads}{\code{cleanCounts} param; minimum number of reads to keep
a gene (default: 100)}

\item{min.lib.size}{\code{cleanCounts} param; minimum number of counts a
pixel needs to keep (default: 100)}

\item{min.detected}{\code{cleanCounts} param; minimum number of pixels a gene
needs to have been detected in to keep (default: 1)}

\item{ODgenes}{Boolean to use \code{getOverdispersedGenes} for the corpus
genes (default: TRUE)}

\item{nTopOD}{number of top overdispersed genes to use. int (default: NA)}

\item{od.genes.alpha}{alpha parameter for \code{getOverdispersedGenes}.
Higher = less stringent and more overdispersed genes returned (default: 0.05)}

\item{gam.k}{gam.k parameter for \code{getOverdispersedGenes}. Dimension
of the "basis" functions in the GAM used to fit, higher = "smoother"
(default: 5)}
}
\value{
A list that contains
\itemize{
\item corpus: (pixels x genes) matrix of the counts of the selected genes
\item slm: slam::as.simple_triplet_matrix(corpus); required format for topicmodels::LDA input
\item positions: matrix of x and y coordinates of pixels rownames = pixels, colnames = "x", "y"
}
}
\description{
Takes pixel (row) x gene (columns) mtx and filters out poor genes
and pixels. Then selects for genes to be included in corpus.
If the pixel IDs are made up of their positions in "XxY" these
can be extracted as the pixel position coordinates.\preformatted{         Order of filtering options:
         1. `cleanCounts` to remove poor pixels and genes
         2. Selection to use specific genes only
         3. Remove top expressed genes in matrix
         4. Remove specific genes based on grepl pattern matching
         5. Remove genes that appear in more than a percentage of pixels
         6. Use the overdispersed genes computed from the remaining genes
            after filtering steps 1-5 (if selected)
}
}
