% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reduceTheta}
\alias{reduceTheta}
\title{Function to reduce theta matrices down to the top X cell-types in each
pixel.}
\usage{
reduceTheta(theta, top = 3)
}
\arguments{
\item{theta}{pixel (rows) by cell-types (columns) distribution matrix. Each row
is the cell-type composition for a given pixel}

\item{top}{Seelct number of top cell-types in each pixel to keep (default: 3)}
}
\value{
A filtered pixel (rows) by cell-types (columns) distribution matrix.
}
\description{
The cell-types with the top X highest proportions are kept in each
pixel and the rest are set to 0. Then renormalizes the pixel proportions to sum to 1.
Cell-types that result in 0 in all pixels after this filtering step are removed.
}
