% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{restrictCorpus}
\alias{restrictCorpus}
\title{Restrict to informative words (genes) for topic modeling}
\usage{
restrictCorpus(counts, t = 0.9, alpha = 0.05, plot = FALSE, verbose = TRUE)
}
\description{
Restrict to informative words (genes) for topic modeling
}
