% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{scale0_1}
\alias{scale0_1}
\title{Helper function to scale values to 0-1 range relative to each other. For use
with \code{lsatPairs}}
\usage{
scale0_1(x)
}
\arguments{
\item{x}{vector or matrix}
}
\value{
vector or matrix with values adjusted to 0-1 scale
}
\description{
Helper function to scale values to 0-1 range relative to each other. For use
with \code{lsatPairs}
}
