% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{simulateBregmaSpots}
\alias{simulateBregmaSpots}
\title{Build hash table of simulated spots for each bregma layer of a given
MERFISH experiment.}
\usage{
simulateBregmaSpots(cellCentroidsAndClass, counts, patch_size = 100)
}
\arguments{
\item{cellCentroidsAndClass}{an "annot.table" data.frame from
"mpoa_merfish_clean.RData" that has been filtered for cells (the rows)
of a given merfish experiment and has the following columns:
c('Centroid_X', 'Centroid_Y', 'Bregma', "Cell_class", "Neuron_cluster_ID")}

\item{counts}{cell x gene count matrix for the individual cells in all the
bregmas for which spots will be simulated}

\item{patch_size}{dimension in um to make the simulated spots for a MERFISH
experiment. The Centroid coords are already in um. (default: 100)}
}
\value{
a hash table where each key is a bregma ID (ie [["-0.04]])
    and the returned object is a list that contains
\itemize{
\item bregmaFullDf: the "annot.table" data.frame for the specific bregma with
    a new column "patch_id" indicating which patch a cell is assigned to. Written
    as "xcoord_ycoord". Simulated spots on the edges are dropped so some cells
    are not assigned to a patch and their patch ID is ""
\item cellTypeTable: table of counts of different cell types in each simulated patch
\item patchTotalCells: vector of total cells in each simulated patch
\item cellTypeCount: vector of counts of unique cell types in each simulated patch
\item cellGexp: individual cell x gene counts matrix for the specific cells
    in the bregma
\item patchGexp: patch x gene count matrix; ie the simulated gene counts for
    each simulated patch
}
}
\description{
uses `hash` to create a hash table of each bregma for a given
    input MERFISH experiment. Goal is to generate simulated spots for each bregma
    by partitioning cells into spots based on their Centroid coordinates. The
    size of the simulated spots is chosen as well but default is 100um x 100um.
    Note that for a given experiment/dataset there are multiple bregma layers. 
    Important: "Cell_class" column is where the cell labels will be pulled
    from. These will be used to construct the: "cellTypeTable", where downstream
    in `buildBregmaCorpus`, will be used to make "gtSpotTopics" and "gtCtGenes".
}
