% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{topGenes}
\alias{topGenes}
\title{Returns top n genes of each deconvolved cell-type for a given beta matrix}
\usage{
topGenes(beta, n = 10)
}
\arguments{
\item{beta}{beta matrix (cell-type gene distribution matrix)}

\item{n}{number of top genes for each deconvolved cell-type to return (default: 10)}
}
\value{
a list where each item is a vector of the top genes and their associated probabilities for
a given deconvolved cell-type
}
\description{
For a given beta matrix (cell-type gene distribution matrix),
returns the top n genes based on their probability.
}
