% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{transparentCol}
\alias{transparentCol}
\title{transparent version of color.}
\usage{
transparentCol(color, percent = 50, name = NULL)
}
\arguments{
\item{color}{color name}

\item{percent}{\% transparency (default: 50)}

\item{name}{an optional name for the color (default: NULL)}
}
\description{
adjust transparency of a color
}
