% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{vizGeneCounts}
\alias{vizGeneCounts}
\title{Visualize gene counts for a given gene in the pixels. Can also see group assignment of
spots.}
\usage{
vizGeneCounts(
  df,
  gene,
  groups = NA,
  group_cols = NA,
  winsorize = 0,
  size = 7,
  stroke = 0.5,
  alpha = 1,
  plotTitle = NA,
  showLegend = TRUE
)
}
\arguments{
\item{df}{data.frame where rows are spots and columns must be at least:
"x", "y" for spot positions in space and "gene" column that is counts
of a gene for each spot.}

\item{gene}{column name of the gene counts in df to be visualized}

\item{groups}{colors the spots lines based on a group or cell layer
they belong to. Needs to be a character vector in the order of the spot
rows in df. Ex: c("0", "1", "0", ...)}

\item{group_cols}{color labels for the groups. Ex: c("0" = "gray", "1" = "red")}

\item{winsorize}{Winsorization quantile}

\item{size}{size of the geom_points to plot (default: 7)}

\item{stroke}{thickness of the geom_point lines to help in emphasizing groups
(default: 0.5)}

\item{alpha}{alpha value of colored pixels (default: 1)}

\item{plotTitle}{option to add a title to the plot}

\item{showLegend}{Boolean to show the plot legend}
}
\value{
a plot where each point is a pixel colored by the expression level of the selected gene
}
\description{
Visualize one gene at a time.
}
\examples{
data(mOB)
pos <- mOB$pos
cd <- mOB$counts
counts <- cleanCounts(cd, min.lib.size = 100) 
df <- merge(as.data.frame(pos), as.data.frame(t(as.matrix(counts))), by = 0)
vizGeneCounts(df = df, gene = "Sox11",
    size = 3, stroke = 0.1, plotTitle = "Sox11",
    winsorize = 0.05, showLegend = TRUE)

}
