% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{vizTopicClusters}
\alias{vizTopicClusters}
\title{Visualize proportions of topic clusters}
\usage{
vizTopicClusters(
  theta,
  pos,
  topicOrder,
  clusters,
  groups = NA,
  group_cols = NA,
  sharedCol = FALSE,
  r = 40,
  lwd = 0.5,
  plotTitle = NA
)
}
\arguments{
\item{theta}{document topic proportion matrix}

\item{pos}{position of documents, x and y columns}

\item{topicOrder}{order of topics based on dendrogram; a numeric vector}

\item{clusters}{factor of the color (topic cluster) each cluster is assigned to
In this case, the levels should be colors. In `vizAllTopics`,
clusters is "cluster_cols" and can just be a vector of colors.}

\item{groups}{color spot piecharts based on a group or cell layer they belong to.
Needs to be a character vector. Ex: c("0", "1", "0", ...).}

\item{group_cols}{color labels for the groups. Ex: c("0" = "gray", "1" = "red")}

\item{r}{= radius of the circles. Adjust based on size of spots.
40 = merfish; 0.4 = mOB}

\item{lwd}{= width of lines of the pie charts}
}
\description{
Visualize proportions of topic clusters
}
