% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{where.is.knee}
\alias{where.is.knee}
\title{Calculates an approximation of the second derivative of a set of points}
\usage{
where.is.knee(points = NULL)
}
\arguments{
\item{points}{vector of perplexity values ordered by the corresponding LDA model K}
}
\value{
the best K (i.e. inflection point of the perplexities)
}
\description{
Calculates an approximation of the second derivative of a set of points.
The maximum second derivative will be a good choice for the inflection point (the elbow or knee).
From https://rdrr.io/cran/gama/src/R/bestk.R
https://stackoverflow.com/questions/2018178/finding-the-best-trade-off-point-on-a-curve
https://raghavan.usc.edu/papers/kneedle-simplex11.pdf (Finding a "Kneedle" in a Haystack:
Detecting Knee Points in System Behavior)
}
