% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{buildLDAobject}
\alias{buildLDAobject}
\title{Wrapper to extract beta (cell type-gene distribution matrix),
theta (pixel cell-type distribution) for individual cell-types and aggregated cell-type-clusters
for an LDA model in \code{fitLDA} output list.}
\usage{
buildLDAobject(
  LDAmodel,
  corpus = NULL,
  clustering = "ward.D",
  dynamic = "hybrid",
  deepSplit = 4,
  colorScheme = "rainbow",
  plot = TRUE
)
}
\arguments{
\item{LDAmodel}{LDA model from fitLDA}

\item{corpus}{If corpus is NULL, then it will use the original corpus that
the model was fitted to. Otherwise, compute deconvolved topics from this
new corpus. Needs to be pixels x genes and nonnegative integer counts.
Each row needs at least 1 nonzero entry (default: NULL)}

\item{clustering}{Clustering agglomeration method to be used (default: ward.D)}

\item{dynamic}{Dynamic tree cutting method to be used (default: hybrid)}

\item{deepSplit}{parameter for \code{clusterTopics} for dynamic tree splitting
when aggregating cell-types (default: 4)}

\item{colorScheme}{color scheme for generating colors assigned to cell-type
clusters for visualizing. Either "rainbow" or "ggplot" (default: "rainbow")}

\item{plot}{Boolean for plotting}
}
\value{
A list that contains
\itemize{
\item beta: cell-type (rows) by gene (columns) distribution matrix.
Each row is a probability distribution of a cell-type expressing each gene
in the corpus
\item theta: pixel (rows) by cell-type (columns) distribution matrix. Each row
is the cell-type composition for a given pixel
\item clusters: factor of the cell-types (names) and their assigned cluster (levels)
\item dendro: dendrogram of the clusters. Returned from \code{stats::hclust()} in \code{clusterTopics}
\item cols: factor of colors for each cell-type where colors correspond to their assigned cluster
\item betaCombn: cell-type (rows) by gene (columns) distribution matrix for combined cell-type-clusters
\item thetaCombn: pixel (rows) by cell-type (columns) distribution matrix for aggregated cell-type-clusters
\item clustCols: factor of colors for each cell-type-cluster
\item k: number of cell-types K of the model
}
}
\description{
Wrapper that combines the functions \code{getBetaTheta}, \code{clusterTopics},
\code{combineTopics} and slots of the topicmodels::LDA object to return a list
that contains the most relevant components of a given LDA model for ease
of analysis and visualization.
}
