% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{buildLDAobject}
\alias{buildLDAobject}
\title{Wrapper to extract beta (topic-gene distribution matrix),
theta (spot-topic distribution) for individual topic and combined topic-clusters
for an LDA model in \code{fitLDA} output list.}
\usage{
buildLDAobject(LDAmodel, deepSplit = 4, colorScheme = "rainbow")
}
\arguments{
\item{LDAmodel}{LDA model from fitLDA}

\item{deepSplit}{parameter for \code{clusterTopics} for dynamic tree splitting
when clustering topics (default: 4)}

\item{colorScheme}{color scheme for generating colors assigned to topic
clusters for visualizing. Either "rainbow" or "ggplot" (default: "rainbow")}
}
\value{
A list that contains
\itemize{
\item beta: topic (rows) by gene (columns) distribution matrix.
Each row is a probability distribution of a topic expressing each gene
in the corpus
\item theta: spot (rows) by topics (columns) distribution matrix. Each row
is the topic composition for a given spot
\item topicFreq: overall proportion of each topic in the entire corpus of
spots
\item clusters: factor of the topics (names) and their assigned cluster (levels)
\item dendro: dendrogram of the clusters. Returned from \code{stats::hclust()} in \code{clusterTopics}
\item cols: factor of colors for each topic where colors correspond to their assigned cluster
\item betaCombn: topic (rows) by gene (columns) distribution matrix for combined topic-clusters
\item thetaCombn: spot (rows) by topic (columns) distribution matrix for combined topic-clusters
\item clustCols: factor of colors for each topic-cluster
\item k: number of topics K of the model
}
}
\description{
Wrapper that combines the functions \code{getBetaTheta}, \code{clusterTopics},
\code{combineTopics} and slots of the topicmodels::LDA object to return a list
that contains the most relevant components of a given LDA model for ease
of analysis and visualization.
}
