% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{clusterTopics}
\alias{clusterTopics}
\title{Aggregate cell-types together using dynamic tree cutting.}
\usage{
clusterTopics(
  beta,
  clustering = "ward.D",
  dynamic = "hybrid",
  deepSplit = 4,
  plot = TRUE
)
}
\arguments{
\item{beta}{Beta matrix (cell-type gene distribution matrix)}

\item{clustering}{Clustering agglomeration method to be used (default: ward.D)}

\item{dynamic}{Dynamic tree cutting method to be used (default: hybrid)}

\item{deepSplit}{Dynamic tree cutting sensitivity parameter (default: 4)}

\item{plot}{Boolean for plotting}
}
\value{
A list that contains
\itemize{
\item order: vector of the dendrogram index order for the cell-types
\item clusters: factor of the cell-types (names) and their assigned cluster (levels)
\item dendro: dendrogram of the clusters
}
}
\description{
Aggregate cell-types together using dynamic tree cutting.
}
