% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{combineTopics}
\alias{combineTopics}
\title{combine topics within same cluster.
the term beta distriubtions are added
or the theta document proportions are added}
\usage{
combineTopics(mtx, clusters, mtxType)
}
\description{
mtx = needs to be topic (row) x feature (term or document) (column) format
clusters = factor of the topics (names) and their assigned cluster (levels)
}
\details{
mtxType = either "b" (beta) or "t" (theta). Affects the adjustment to the
          combined topic vectors. "b" divides summed topic vectors by number
          of combined topics.
}
