% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{correlationPlot}
\alias{correlationPlot}
\title{Generate heatmap of correlations}
\usage{
correlationPlot(mat, colLabs = NA, rowLabs = NA, title = NA)
}
\arguments{
\item{mat}{matrix with correlation values from -1 to 1}

\item{colLabs}{x-axis label for plot. These are the columns of the matrix, or specifically m2 from getCorrMtx. (default: NULL)}

\item{rowLabs}{y-axis label for plot. These are the rows of the matrix, or specifically m1 from getCorrMtx. (default: NULL)}

\item{title}{title of the plot. (default: NULL)}
}
\description{
Visualize the correlations between topics stored in a matrix, typically one
returned via \code{getCorrMtx()}. This function uses ggplot2::geom_tile.
}
