% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fitLDA}
\alias{fitLDA}
\title{Find the optimal number of cell-types K for the LDA model}
\usage{
fitLDA(
  counts,
  Ks = seq(2, 10, by = 2),
  seed = 0,
  testSize = NULL,
  perc.rare.thresh = 0.05,
  ncores = parallel::detectCores(logical = TRUE) - 1,
  plot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{counts}{Gene expression counts with pixels as rows and genes as columns}

\item{Ks}{vector of K parameters, or number of cell-types, to fit models with}

\item{seed}{Random seed}

\item{testSize}{fraction of pixels to set aside for test corpus when computing perplexity (default: NULL)
Either NULL or decimal between 0 and 1.}

\item{perc.rare.thresh}{the number of deconvolved cell-types with mean pixel proportion below this fraction used to assess
performance of fitted models for each K. Recorded for each K. (default: 0.05)}

\item{ncores}{Number of cores for parallelization}

\item{plot}{Boolean for plotting (default: TRUE)}

\item{verbose}{Boolean for verbosity (default: TRUE)}
}
\value{
A list that contains
\itemize{
\item models: each fitted LDA model for a given K
\item kneedOptK: the optimal K based on Kneed algorithm
\item minOptK: the optimal K based on minimum
\item ctPropOptK: Suggested upper bound on K. K in which number of returned cell-types
with mean proportion < perc.rare.thresh starts to increase steadily.
\item numRare: number of cell-types with mean pixel proportion < perc.rare.thresh for each K
\item perplexities: perplexity scores for each model
\item fitCorpus: the corpus that was used to fit each model
\item testCorpus: the corpus used to compute model perplexity. If testSize = NULL then same as fitCorpus.
}
}
\description{
The input for topicmodels::LDA needs to be a
slam::as.simple_triplet_matrix (docs x words). Access a given model in
the returned list via: lda$models$k. The models are objects from
the R package "topicmodels". The LDA models have slots with additional
information.
}
