% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getBetaTheta}
\alias{getBetaTheta}
\title{Pull out cell-type proportions across pixels (theta) and
cell-type gene probabilities (beta) matrices from fitted LDA models from fitLDA}
\usage{
getBetaTheta(lda, corpus)
}
\arguments{
\item{lda}{an LDA model from \code{topicmodels}. From list of models returned by
fitLDA}

\item{corpus}{Gene expression counts with pixels as rows and genes as columns
for which to get predicted cell-type proportions and cell-type gene expression profiles.
Typically same one that was used to fit the LDA model.}
}
\value{
A list that contains
\itemize{
\item beta: cell-type (rows) by gene (columns) distribution matrix.
Each row is a probability distribution of a cell-type expressing each gene
in the corpus
\item theta: pixel (rows) by cell-types (columns) distribution matrix. Each row
is the cell-type composition for a given pixel
}
}
\description{
Pull out cell-type proportions across pixels (theta) and
cell-type gene probabilities (beta) matrices from fitted LDA models from fitLDA
}
