% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getBetaTheta}
\alias{getBetaTheta}
\title{Pull out cell-type proportions across pixels (theta) and
cell-type gene probabilities (beta) matrices from fitted LDA models from fitLDA}
\usage{
getBetaTheta(lda)
}
\arguments{
\item{lda}{an LDA model from \code{topicmodels}. From list of models returned by
fitLDA}
}
\value{
A list that contains
\itemize{
\item beta: cell-type (rows) by gene (columns) distribution matrix.
Each row is a probability distribution of a cell-type expressing each gene
in the corpus
\item theta: pixel (rows) by cell-types (columns) distribution matrix. Each row
is the cell-type composition for a given pixel
\item topicFreq: overall proportion of each cell-type in the entire corpus of
pixels
}
}
\description{
Pull out cell-type proportions across pixels (theta) and
cell-type gene probabilities (beta) matrices from fitted LDA models from fitLDA
}
