% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getBetaTheta}
\alias{getBetaTheta}
\title{Pull out cell-type proportions across pixels (theta) and
cell-type gene probabilities (beta) matrices from fitted LDA models from fitLDA}
\usage{
getBetaTheta(
  lda,
  corpus = NULL,
  perc.filt = 0.05,
  betaScale = 1,
  verbose = TRUE
)
}
\arguments{
\item{lda}{an LDA model from "topicmodels" R package. From list of models returned by
fitLDA}

\item{corpus}{If corpus is NULL, then it will use the original corpus that
the model was fitted to. Otherwise, compute deconvolved topics from this
new corpus. Needs to be pixels x genes and nonnegative integer counts.
Each row needs at least 1 nonzero entry (default: NULL)}

\item{perc.filt}{proportion threshold to remove cell-types in pixels (default: 0.05)}

\item{betaScale}{factor to scale the predicted cell-type gene expression profiles (default: 1)}

\item{verbose}{Boolean for verbosity (default: TRUE)}
}
\value{
A list that contains
\itemize{
\item beta: cell-type (rows) by gene (columns) distribution matrix.
Each row is a probability distribution of a cell-type expressing each gene
in the corpus
\item theta: pixel (rows) by cell-types (columns) distribution matrix. Each row
is the cell-type composition for a given pixel
}
}
\description{
Pull out cell-type proportions across pixels (theta) and
cell-type gene probabilities (beta) matrices from fitted LDA models from fitLDA
}
\examples{
data(mOB)
pos <- mOB$pos
cd <- mOB$counts
counts <- cleanCounts(cd, min.lib.size = 100)
corpus <- restrictCorpus(counts, removeAbove=1.0, removeBelow = 0.05)
ldas <- fitLDA(t(as.matrix(corpus)), Ks = seq(2,6))
optLDA <- optimalModel(models = ldas, opt = 6)
results <- getBetaTheta(optLDA, perc.filt = 0.05, betaScale = 1000)
head(results$theta)
head(results$beta)

}
