% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getCorrMtx}
\alias{getCorrMtx}
\title{Find Pearson's correlations between topics (cell-types) with respect to their
proportions across documents (pixels), i.e. thetas, or gene probabilities,
i.e. betas.}
\usage{
getCorrMtx(m1, m2, type, thresh = NULL, verbose = TRUE)
}
\arguments{
\item{m1}{first matrix}

\item{m2}{second matrix}

\item{type}{must be either "t" (theta; cell-type proportions across pixels) or "b" (beta; cell-type gene expression profiles)}

\item{thresh}{if comparing betas, use to compare genes above this probability (e.g., expression level).
NULL or 0 < numeric < 1.0 (default: NULL)}

\item{verbose}{control the verbosity (default: TRUE)}
}
\value{
matrix of Pearson's correlations; m1 (rows) by m2 (cols)
}
\description{
Find Pearson's correlations between topics (cell-types) with respect to their
proportions across documents (pixels), i.e. thetas, or gene probabilities,
i.e. betas.
}
