% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{optimalModel}
\alias{optimalModel}
\title{Get the optimal LDA model}
\usage{
optimalModel(models, opt)
}
\arguments{
\item{models}{list returned from fitLDA}

\item{opt}{either "kneed" (kOpt1) or "min" (kOpt2), or designate a specific K}
}
\value{
optimal LDA model fitted to the K based on `opt`
}
\description{
Get the optimal LDA model
}
