% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{optimalModel}
\alias{optimalModel}
\title{Get the optimal LDA model}
\usage{
optimalModel(models, opt)
}
\arguments{
\item{models}{list returned from fitLDA}

\item{opt}{either "kneed" (kOpt1) or "min" (kOpt2), or designate a specific K.
"kneed" = K vs perplexity inflection point.
"min" = K corresponding to minimum perplexity
"proportion" = K vs number of cell-type with mean proportion < 5\% inflection point}
}
\value{
optimal LDA model fitted to the K based on opt parameter
}
\description{
Get the optimal LDA model
}
\examples{
data(mOB)
pos <- mOB$pos
cd <- mOB$counts
counts <- cleanCounts(cd, min.lib.size = 100)
corpus <- restrictCorpus(counts, removeAbove=1.0, removeBelow = 0.05)
ldas <- fitLDA(t(as.matrix(corpus)), Ks = seq(2,4), ncores=7)
optLDA <- optimalModel(models = ldas, opt = "min")

}
