% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{perplexityPlot}
\alias{perplexityPlot}
\title{Plot the perplexity and rare cell-types versus fitted Ks}
\usage{
perplexityPlot(models, corpus = NULL, perc.rare.thresh = 0.05)
}
\arguments{
\item{models}{models output from fitLDA()}

\item{corpus}{If corpus is NULL, then it will use the original corpus that
the model was fitted to. Otherwise, compute deconvolved topics from this
new corpus. Needs to be pixels x genes and nonnegative integer counts.
Each row needs at least 1 nonzero entry (default: NULL)}

\item{perc.rare.thresh}{the number of deconvolved cell-types with mean pixel proportion below this fraction used to assess
performance of fitted models for each K. Recorded for each K. (default: 0.05)}
}
\description{
the same plot returned by fitLDA() but now callable as a
separate function.
}
