% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{topGenes}
\alias{topGenes}
\title{Returns top n genes of each deconvolved cell-type for a given beta matrix}
\usage{
topGenes(beta, n = 10)
}
\arguments{
\item{beta}{beta matrix (cell-type gene distribution matrix)}

\item{n}{number of top genes for each deconvolved cell-type to return (default: 10)}
}
\value{
a list where each item is a vector of the top genes and their associated probabilities for
a given deconvolved cell-type
}
\description{
For a given beta matrix (cell-type gene distribution matrix),
returns the top n genes based on their probability.
}
\examples{
data(mOB)
pos <- mOB$pos
cd <- mOB$counts
counts <- cleanCounts(cd, min.lib.size = 100)
corpus <- restrictCorpus(counts, removeAbove=1.0, removeBelow = 0.05)
ldas <- fitLDA(t(as.matrix(corpus)), Ks = seq(2,6), ncores=7)
optLDA <- optimalModel(models = ldas, opt = 6)
results <- getBetaTheta(optLDA, perc.filt = 0.05, betaScale = 1000)
deconGexp <- results$beta
genes <- topGenes(deconGexp)

}
