% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{vizAllTopics}
\alias{vizAllTopics}
\title{Visualize topic proportions across spots with `scatterpie`}
\usage{
vizAllTopics(
  theta,
  pos,
  topicOrder = seq(ncol(theta)),
  cluster_cols = rainbow(ncol(theta)),
  groups = NA,
  group_cols = NA,
  r = 1,
  lwd = 0.5,
  showLegend = TRUE,
  plotTitle = NA
)
}
\arguments{
\item{theta}{document x topic proportion matrix}

\item{pos}{position of documents, x and y columns}

\item{topicOrder}{order of topics based on dendrogram; a numeric vector
from: (clusterTopics$order)}

\item{cluster_cols}{vector of colors for each of the topics
can use factor of clusters for each topic (via clusterTopics$clusters)
as long as the cluster levels/values have been converted to colors.
This gets reordered wrt the topicOrder fyi.}

\item{groups}{color spot piecharts based on a group or cell layer they belong to.
Needs to be a character vector. Ex: c("0", "1", "0", ...).}

\item{group_cols}{color labels for the groups. Ex: c("0" = "gray", "1" = "red")}

\item{r}{= radius of the circles. Adjust based on size of spots.}

\item{lwd}{= width of lines of the pie charts}
}
\description{
Visualize topic proportions across spots with `scatterpie`
}
