% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{vizAllTopics}
\alias{vizAllTopics}
\title{Visualize all topic proportions across pixels with \code{scatterpie}}
\usage{
vizAllTopics(
  theta,
  pos,
  topicOrder = seq(ncol(theta)),
  topicCols = rainbow(ncol(theta)),
  groups = NA,
  group_cols = NA,
  r = max(0.4, max(pos)/nrow(pos) * 4),
  lwd = 0.5,
  showLegend = TRUE,
  plotTitle = NA,
  overlay = NA
)
}
\arguments{
\item{theta}{document (pixel) x cell-type proportion matrix}

\item{pos}{position of pixels, as data.frame with \code{x} and \code{y} columns}

\item{topicOrder}{order of topics in theta to visualize as a numeric vector
and same length as topicCols (default: seq(ncol(theta)))}

\item{topicCols}{Vector of colors for each of the cell-types to be visualized.
Same length and order as topicOrder (default: rainbow(ncol(theta)))}

\item{groups}{Indicates color of the scatterpie strokes with the goal of coloring them
by their assigned group. This can be a vector or factor indicating the group of each
scatterpie. Needs to be in the same order as the pixel rows in "theta" (default: NA)}

\item{group_cols}{Color labels for the groups. Can be a vector or factor. (default: NA)}

\item{r}{Radius of the scatterpie circles. Adjust based on positions of pixels (default: max(0.4, max(pos)/nrow(pos)*4))}

\item{lwd}{Width of lines of the pie charts. Increasing helps visualize
group_cols if being used.}

\item{showLegend}{Boolean to show the legend indicating cell-types and their color}

\item{plotTitle}{add title to the resulting plot (default: NA)}

\item{overlay}{plot the scatterpies on top of a raster image of the H&E tissue
(default: NA)}
}
\description{
Note: visualizes all cell-types in theta at once (could be
individual cell-types or cell-type-clusters) so for accuracy of the proportions
of each cell-type in a pixel, the row (pixel) should sum to 1.
}
