% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{vizTopic}
\alias{vizTopic}
\title{Visualize pixel proportions of a single cell-type.}
\usage{
vizTopic(
  theta,
  pos,
  topic,
  groups = NA,
  group_cols = NA,
  size = 2,
  stroke = 0.3,
  alpha = 1,
  low = "white",
  high = "red",
  plotTitle = NA,
  showLegend = TRUE
)
}
\arguments{
\item{theta}{document (pixel) x cell-type proportion matrix}

\item{pos}{position of pixels, as data.frame with \code{x} and \code{y} columns}

\item{topic}{the index of the topic}

\item{groups}{colors the pixel border lines based on a group or cell layer
they belong to. Needs to be a character or named vector of assigned groups for each pixel
Ex: c("0", "1", "0", ...)}

\item{group_cols}{color labels for the groups. Ex: c("0" = "gray", "1" = "red")}

\item{size}{size of the geom_points to plot (default: 2)}

\item{stroke}{thickness of the geom_point lines to help in emphasizing groups
(default: 0.5)}

\item{alpha}{alpha value of colored pixels (default: 1)}

\item{low}{sets the color for the low end of the topic proportion color scale (default: "white")}

\item{high}{sets the color the the high end of the topic proportion color scale (default: "red")}

\item{plotTitle}{option to add a title to the plot (character)}

\item{showLegend}{Boolean to show the plot legend}
}
\value{
a plot where each point is a pixel colored by the proportion of the selected cell-type
}
\description{
Visualize the pixel proportions of a single topic.
}
\examples{
data(mOB)
pos <- mOB$pos
cd <- mOB$counts
counts <- cleanCounts(cd, min.lib.size = 100)
corpus <- restrictCorpus(counts, removeAbove=1.0, removeBelow = 0.05)
ldas <- fitLDA(t(as.matrix(corpus)), Ks = 3)
optLDA <- optimalModel(models = ldas, opt = 3)
results <- getBetaTheta(optLDA, perc.filt = 0.05, betaScale = 1000)
deconProp <- results$theta
vizTopic(theta = deconProp, pos = pos, topic = "3", plotTitle = "X3",
    size = 5, stroke = 1, alpha = 0.5, low = "white", high = "red")

}
